refactor: quick action compare/preview
